Change int to float in Drawable Move

+++ src/gui/drawable.cpp
@@ -24,7 +24,7 @@ void Drawable::Disable(bool disable) {
     body_.setFillColor(sf::Color(255 - DISABLE * disable, 255 - DISABLE * disable, 255 - DISABLE * disable, 255));
 }
 
-void Drawable::Move(int d_x, int d_y) {
+void Drawable::Move(float d_x, float d_y) {
     auto current_pos = body_.getPosition();
-    body_.setPosition(sf::Vector2f((float)d_x / WINDOW_WIDTH + current_pos.x, (float)d_y / WINDOW_HEIGHT + current_pos.y));
+    body_.setPosition(sf::Vector2f(d_x / WINDOW_WIDTH + current_pos.x, d_y / WINDOW_HEIGHT + current_pos.y));
 }
